DEV-136: Add delete product functionality
