+++ app/components/organisms/participant-products-upload/index.tsx
@@ -91,6 +91,7 @@ export function ParticipantProductsUpload({
 				onOpenChange={handleToggleProductModal}
 				onClose={() => {
 					setShowProductModal(false);
+					resetModal();
 				}}
 				uploadedImageUrl={uploadedImageUrl}
 				setUploadedImageUrl={setUploadedImageUrl}
